server/requirements.txt
arcade==2.6.17
asgiref==3.8.1
attrs==22.2.0
# cffi==1.15.1
Django
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
Markdown==3.6
numpy==2.0.0
pandas==2.2.2
# Pillow==9.3.0
pycparser==2.21
# pygame==2.3.0
# pyglet==2.0.dev23
# pymunk==6.4.0
python-dateutil==2.9.0.post0
pytiled-parser==2.2.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.5.0
tzdata==2024.1
